+++ source/menuscreenshot.c
@@ -242,7 +242,7 @@ void menuScreenshot()
         str_hint_asel,
     };
 
-    s8 cursor = 0;
+    s32 cursor = 0;
     u8 selection = menuInputs(&cursor, 2, (struct InputIDs) {1,0,0}, 1, 1, 1, (sizeof(ptr_array) / sizeof(ptr_array[0])), ptr_array);
 
     switch(selection)
